Remove most of crypto and move out of search path
We only use keccak out of cncrypto, so this removes all the other junk,
and minimally modifies the keccak implementation to build in ethyl
without needing other files (and updates the keccak signature to match
oxen-core's, which changed the last argument from int to size_t).

This also puts it under a src/ethyl-keccak subdirectory so that when we
are building with cncrypto from oxen-core, we won't find the local
version in our include path and instead get the appropriate headers via
oxen-core when building with oxen-core's cncrypto.

+++ src/utils.cpp
@@ -62,7 +62,7 @@ Bytes32 utils::hashHex(std::string_view hex) {
 
 Bytes32 utils::hashBytesPtr(const void *bytes, size_t size) {
   Bytes32 result;
-  keccak(reinterpret_cast<const uint8_t*>(bytes), size, result.data(), static_cast<int>(result.max_size()));
+  keccak(reinterpret_cast<const uint8_t*>(bytes), size, result.data(), result.max_size());
   return result;
 }
